rename package directory to lablink_client
