+++ lablink_client/update_inuse_status.py
@@ -13,8 +13,8 @@ except ImportError as e:
     )
     raise e
 
-from tutorial_vm.database import SpannerDatabase
-from tutorial_vm.logging_utils import CloudAndConsoleLogger
+from lablink_client.database import SpannerDatabase
+from lablink_client.logging_utils import CloudAndConsoleLogger
 
 
 cnc_logger = CloudAndConsoleLogger(
